% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cfin_CV_CVI_GOM}
\alias{Cfin_CV_CVI_GOM}
\title{C. finmarchicus monthly vertical distribution in the Gulf oF Maine}
\format{
GAM object:
}
\source{
Jonhson, O. et al. in prep.
}
\usage{
Cfin_CV_CVI_GOM
}
\description{
A GAM object to provided 3D distribution of Calanus finmarhcicus in the Gulf of Maine
}
\keyword{datasets}
