% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_vertical_run_model.R
\name{execute_bioenergy}
\alias{execute_bioenergy}
\title{Transform Calanus biomass into Enet for NARW}
\usage{
execute_bioenergy(
  dat,
  cfin = "Z_Cf.Cg4.6_mgm2",
  chyp = "Z_Ch4.6_mgm2",
  psca = "Z_Pcal.Micro.Scol_mgm2",
  temo = "Z_Tem.Eury_mgm2",
  cfinZ = NULL,
  chypZ = NULL,
  pscaZ = NULL,
  temoZ = NULL,
  units = "mg",
  ID = "Label",
  change_150m = T,
  state = "pregnant",
  param = "mean"
)
}
\arguments{
\item{dat}{dataframe to import}

\item{cfin}{name of the column (quoted) with Cfin biomass}

\item{chyp}{name of the column (quoted) with Chyp biomass}

\item{psca}{name of the column (quoted) with Pseudocalanus spp. biomass}

\item{temo}{name of the column (quoted) with temora biomass}

\item{cfinZ}{default to NULL, optionnaly if the vertical distribution doesn't not need to be explored, name of the column for depth of the maximum of Cfin biomass.}

\item{chypZ}{default to NULL, optionnaly if the vertical distribution doesn't not need to be explored, name of the column for depth of the maximum of Chyp biomass.}

\item{pscaZ}{default to NULL, optionnaly if the vertical distribution doesn't not need to be explored, name of the column for depth of the maximum of pseudocalanus biomass.}

\item{temoZ}{default to NULL, optionnaly if the vertical distribution doesn't not need to be explored, name of the column for depth of the maximum of temora biomass.}

\item{units}{units of the biomass, one of "ug", "mg", "g", "kJ", default to mg}

\item{ID}{name of the column (quoted) for the unique ID of samples}

\item{change_150m}{does the NARW behavior change when it dives below 150m. Default to T  see details}

\item{state}{decide to extract results for "resting", "pregnant" or "lactating" females. default to "pregnant". At the moment does not accommodate more that one state.}

\item{param}{decide to extract results for "min", "mean" or "max" enet. default to "mean". At the moment does not accommodate more that one state.}
}
\value{
the same data frame as the input, with added columns for NARW bioenergy.
}
\description{
Transform Calanus biomass into Enet for NARW
}
\details{
Z columns applies for Nicolas Lecorre outputs. I don't see other possible usage.
The parameter change_150m is exploratory. In Gavrilchuk et al. and Lehoux et al. the whale change its behaviour below 150m , the dive has a maximum duration and thus time spent feedding decrease with feeding depth. If FALSE, then feeding duration decrease with depth for all depth. I strongly suggest using TRUE except for exploration.
}
\examples{
#TO DO
}
