% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_vertical_run_model.R
\name{kj_df}
\alias{kj_df}
\title{Apply vertical distribution to integrated energy}
\usage{
kj_df(datkj, ID)
}
\arguments{
\item{datkj}{dataframe}

\item{ID}{name of the column for unique identifier.}
}
\value{
a dataframe with Zsampled for 3D distribution
}
\description{
Apply vertical distribution to integrated energy
}
\examples{
#to do
}
