% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcalanus_c_ess}
\alias{mcalanus_c_ess}
\title{Calanus monthly vertical distribution on the Eastern Scotian Shelf (ESS).}
\format{
GAM object:
}
\source{
Jonhson, O. et al. in prep.
}
\usage{
mcalanus_c_ess
}
\description{
A GAM object to provided 3D distribution of Calanus spp. on the ESS
}
\keyword{datasets}
