% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcalanus_c_gsl}
\alias{mcalanus_c_gsl}
\title{Calanus monthly vertical distribution in the GSL.}
\format{
GAM object:
}
\source{
\url{https://academic.oup.com/plankt/article/41/5/667/5532756?login=true}
}
\usage{
mcalanus_c_gsl
}
\description{
A GAM object to provided 3D distribution of Calanus spp.
}
\keyword{datasets}
