% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcfin_glac_c_gsl}
\alias{mcfin_glac_c_gsl}
\title{Calanus finmarchicus/glacialis monthly vertical distribution in the GSL.}
\format{
GAM object:
}
\source{
\url{https://academic.oup.com/plankt/article/41/5/667/5532756?login=true}
}
\usage{
mcfin_glac_c_gsl
}
\description{
A GAM object to provided 3D distribution of Calanus finmarchicus/glacialis
}
\keyword{datasets}
