% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcfin_glac_c_gsl_fall}
\alias{mcfin_glac_c_gsl_fall}
\title{Calanus finmarchicus/glacialis vertical distribution in the GSL during fall.}
\format{
GAM object:
}
\source{
\url{https://publications.gc.ca/collections/collection_2020/mpo-dfo/fs70-5/Fs70-5-2020-033-eng.pdf}
}
\usage{
mcfin_glac_c_gsl_fall
}
\description{
A GAM object to provided 3D distribution of Calanus finmarchicus/glacialis
}
\keyword{datasets}
