% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mchyp_c_ess}
\alias{mchyp_c_ess}
\title{C. hyperboreus monthly vertical distribution on the Eastern Scotian Shelf (ESS).}
\format{
GAM object:
}
\source{
\url{https://academic.oup.com/plankt/article/41/5/667/5532756?login=true}
}
\usage{
mchyp_c_ess
}
\description{
A GAM object to provided 3D distribution of C.hyperboreus. on the ESS
}
\keyword{datasets}
