% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mchyp_c_gsl_fall}
\alias{mchyp_c_gsl_fall}
\title{Calanus hyperboreus vertical distribution in the GSL during fall.}
\format{
GAM object:
}
\source{
\url{https://publications.gc.ca/collections/collection_2020/mpo-dfo/fs70-5/Fs70-5-2020-033-eng.pdf}
}
\usage{
mchyp_c_gsl_fall
}
\description{
A GAM object to provided 3D distribution of Calanus hyperboreus
}
\keyword{datasets}
