% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mchyp_dev_catN450}
\alias{mchyp_dev_catN450}
\title{Calanus hyperboreus vertical distribution depending on its development index Night}
\format{
GAM object:
}
\source{
Comtois et al. in  prep.
}
\usage{
mchyp_dev_catN450
}
\description{
A GAM object to provided 3D distribution of Calanus hyperboreus
}
\keyword{datasets}
