% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpseudo_c_gsl_N}
\alias{mpseudo_c_gsl_N}
\title{Pseudocalanus spp. vertical distribution in the GSL during the Night.}
\format{
GAM object:
}
\source{
\url{https://publications.gc.ca/collections/collection_2020/mpo-dfo/fs70-5/Fs70-5-2020-033-eng.pdf}
}
\usage{
mpseudo_c_gsl_N
}
\description{
A GAM object to provided 3D distribution of Pseudocalanus spp
}
\keyword{datasets}
