% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtem_c_gsl_D}
\alias{mtem_c_gsl_D}
\title{Temora vertical distribution in the GSL during the Day.}
\format{
GAM object:
}
\source{
\url{https://publications.gc.ca/collections/collection_2020/mpo-dfo/fs70-5/Fs70-5-2020-033-eng.pdf}
}
\usage{
mtem_c_gsl_D
}
\description{
A GAM object to provided 3D distribution of Temora
}
\keyword{datasets}
