% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioenergetic_models.R
\name{NARW_bionerg}
\alias{NARW_bionerg}
\title{Gavrilchuk bioenergetic model}
\usage{
NARW_bionerg(input_file, coeff.capt = 1, mean.ed, sd.ed, change_150m, ID)
}
\arguments{
\item{input_file}{dataframe}

\item{coeff.capt}{depending on species. see lehoux et al. 2020}

\item{mean.ed}{mean energy density of the taxa consumed}

\item{sd.ed}{standard deviation of the energy density of the  taxa consumed}

\item{change_150m}{logicial does the behaviour change below 150m}

\item{ID}{unique ID of sample}
}
\value{
a data frame
}
\description{
Gavrilchuk bioenergetic model
}
\examples{
#todo
}
